% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadc_eds_handler.R
\name{aadc_eds_get}
\alias{aadc_eds_get}
\title{Handler for files downloaded from the Australian Antarctic Data Centre EDS system}
\usage{
aadc_eds_get(data_source)
}
\arguments{
\item{data_source}{tibble: single-row tibble defining a data source, e.g. as returned by \code{bb_source}}
}
\value{
TRUE on success
}
\description{
AADC EDS files have a URL of the form https://data.aad.gov.au/eds/file/wxyz/ or https://data.aad.gov.au/eds/wxyz/download where wxyz is a numeric file identifier. Note that clobber=1 does not currently work with AADC sources (server error, being investigated).
}
\references{
http://data.aad.gov.au
}
