% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{add}
\alias{add}
\title{Add new data sources to a bowerbird configuration}
\usage{
add(cf, source)
}
\arguments{
\item{cf}{data.frame: configuration, as returned by \code{bb_config}}

\item{source}{data.frame: one or more data source definitions, as returned by \code{bb_source}, to add to the configuration}
}
\value{
configuration object
}
\description{
Add new data sources to a bowerbird configuration
}
\examples{
\dontrun{
  cf <- bb_config("/my/file/root") \%>\%
    add(bb_sources("NSIDC SMMR-SSM/I Nasateam sea ice concentration"))
}
}
\seealso{
\code{\link{bb_source}} \code{\link{bb_config}}
}
