% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amps_handler.R
\name{amps_get}
\alias{amps_get}
\title{Handler for AMPS data (Antarctic mesoscale prediction system)}
\usage{
amps_get(data_source)
}
\arguments{
\item{data_source}{tibble: single-row tibble defining a data source, e.g. as returned by \code{bb_source}}
}
\value{
TRUE on success
}
\description{
Handler for AMPS data (Antarctic mesoscale prediction system)
}
\references{
http://www2.mmm.ucar.edu/rt/amps/
}
