% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bb_add}
\alias{bb_add}
\title{Add new data sources to a bowerbird configuration}
\usage{
bb_add(config, source)
}
\arguments{
\item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config})}

\item{source}{data.frame: one or more data source definitions, as returned by \code{bb_source}, to add to the configuration}
}
\value{
configuration object
}
\description{
Add new data sources to a bowerbird configuration
}
\examples{
\dontrun{
  cf <- bb_config("/my/file/root") \%>\%
    bb_add(bb_example_sources())
}
}
\seealso{
\code{\link{bb_source}} \code{\link{bb_config}}
}
