% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{bb_add_snapshot}
\alias{bb_add_snapshot}
\title{Add a snapshot to a source}
\usage{
bb_add_snapshot(src, snapshot)
}
\arguments{
\item{src}{data.frame or tibble: a single-row data source (as returned by \code{bb_source})}

\item{snapshot}{data.frame: as returned by \code{\link{bb_get}}}
}
\value{
A modified copy of \code{src}
}
\description{
Experimental! Allows the results of a previous sync run to be added to a source, so that it can be used next time instead of re-indexing the remote site.
}
\examples{
\dontrun{
  ## general usage: first define a source
  src <- bb_source(...)
  ## sync it once to index the remote site. This can be done as a dry run
  result <- bb_get(src, ..., dry_run = TRUE)
  ## add the snapshot to the source
  src <- bb_add_snapshot(src, result)
  ## now the sync can be run again, and the remote site does not need to be re-indexed
  result2 <- bb_get(src, ...)
}

}
