% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bb_attributes}
\alias{bb_attributes}
\title{Returns a configuration object's bowerbird-specific attributes}
\usage{
bb_attributes(cf)
}
\arguments{
\item{cf}{data.frame: configuration, as returned by \code{bb_config}}
}
\value{
named list
}
\description{
Returns a configuration object's bowerbird-specific attributes
}
\examples{
cf <- bb_config(local_file_root="/your/data/directory")
bb_attributes(cf)

}
\seealso{
\code{\link{bb_config}}
}
