% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bb_config}
\alias{bb_config}
\alias{bb_save_config}
\alias{add}
\title{Load or initialize bowerbird configuration}
\usage{
bb_config(file)

bb_save_config(cf, file)

add(cf, source)
}
\arguments{
\item{file}{string: file name of configuration to load. All other parameters are ignored if \code{file} is provided.}

\item{cf}{list: configuration, as returned by \code{bb_config}}

\item{source}{tibble: external data source definition to add to the configuration}
}
\value{
configuration
}
\description{
If \code{file} is provided, load that configuration. Otherwise initialize an empty configuration.
}
\examples{
\dontrun{
  cf <- bb_config() \%>\%
    add(bb_source("seaice_smmr_ssmi_nasateam"))

  ## save to file
  bb_save_config(cf,file="saved_config.json")
  ## load previously saved config
  cf <- bb_config(file="saved_config.json")
}

}
\seealso{
\code{\link{bb_source}}
}
