% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_handler.R
\name{bb_copernicus_cleanup}
\alias{bb_copernicus_cleanup}
\title{Postprocessing: remove redundant Copernicus files}
\usage{
bb_copernicus_cleanup(...)
}
\arguments{
\item{...}{: extra parameters passed automatically by \code{bb_sync}}
}
\value{
a list, with components \code{status} (TRUE on success) and \code{deleted_files} (character vector of paths of files that were deleted)
}
\description{
This function is not intended to be called directly, but rather is specified as a \code{postprocess} option in \code{\link{bb_source}}.
}
\details{
This function will remove files from a Copernicus collection that have been superseded by a more recently-processed version. You might see, for example, the file \code{nrt_global_allsat_phy_l4_20250910_20250910.nc} which is then superseded by \code{nrt_global_allsat_phy_l4_20250910_20250913.nc} at a later date.
}
