% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bb_data_sources}
\alias{bb_data_sources}
\alias{bb_data_sources<-}
\title{Gets or sets a bowerbird configuration object's data sources}
\usage{
bb_data_sources(config)

bb_data_sources(config) <- value
}
\arguments{
\item{config}{bb_config: a bowerbird configuration (as returned by \code{\link{bb_config}})}

\item{value}{data.frame: new data sources to set (e.g. as returned by \code{\link{bb_example_sources}})}
}
\value{
a tibble with columns as specified by \code{\link{bb_source}}
}
\description{
Gets or sets the data sources contained in a bowerbird configuration object.
}
\details{
Note that an assignment along the lines of \code{bb_data_sources(cf) <- new_sources} replaces all of the sources in the configuration with the \code{new_sources}. If you wish to modify the existing sources then read them, modify as needed, and then rewrite the whole lot back into the configuration object.
}
\examples{
## create a configuration and add data sources
cf <- bb_config(local_file_root="/your/data/directory")
cf <- bb_add(cf,bb_example_sources())

## examine the sources contained in cf
bb_data_sources(cf)

## replace the sources with different ones
\dontrun{
bb_data_sources(cf) <- new_sources
}

}
\seealso{
\code{\link{bb_config}}, \code{\link{bb_source}}, \code{\link{bb_example_sources}}
}
