% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_sources.R
\name{bb_example_sources}
\alias{bb_example_sources}
\title{Example bowerbird data sources}
\usage{
bb_example_sources()
}
\value{
tibble
}
\description{
Five example data sources:
\itemize{
  \item "NOAA OI SST V2"
  \item "CMEMS global gridded SSH reprocessed (1993-ongoing)" - a data source that requires a username and password
  \item "Oceandata SeaWiFS Level-3 mapped monthly 9km chl-a" - an example data source that uses the \code{bb_handler_oceandata} method
  \item "Nimbus Ice Edge Points from Nimbus Visible Imagery" - an example data source that uses the \code{bb_handler_earthdata} method
  \item "Bathymetry of Lake Superior" - an example that passes extra flags to the \code{bb_handler_wget} call in order to restrict what is downloaded
}
}
\examples{
\dontrun{
cf <- bb_config("/my/file/root") \%>\%
  bb_add(bb_example_sources()[5,])
bb_sync(cf)
}
}
\references{
See the \code{reference} and \code{citation} field in each row of the returned tibble for references associated with these particular data sources
}
\seealso{
\code{\link{bb_config}} \code{\link{bb_handler_wget}} \code{\link{bb_handler_oceandata}} \code{\link{bb_handler_earthdata}}
}
