% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{bb_fingerprint}
\alias{bb_fingerprint}
\title{Fingerprint the files associated with a data source}
\usage{
bb_fingerprint(config, hash = "sha1")
}
\arguments{
\item{config}{bb_config: configuration as returned by \code{\link{bb_config}}}

\item{hash}{string: algorithm to use to calculate file hashes: "md5", "sha1", or "none". Note that file hashing can be slow for large file collections}
}
\value{
tibble
}
\description{
The \code{bb_fingerprint} function, given a data repository configuration, will return the timestamp of download and hashes of all files associated with its data sources. This is intended as a general helper for tracking data provenance: for all of these files, we have information on where they came from (the data source ID), when they were downloaded, and a hash so that later versions of those files can be compared to detect changes. See also \code{vignette("data_provenance")}.
}
\examples{
\dontrun{
  cf <- bb_config("/my/file/root") \%>\%
    bb_add(bb_example_sources())
  bb_fingerprint(cf)
}

}
\seealso{
\code{vignette("data_provenance")}
}
