% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthdata_stac_handler.R
\name{bb_handler_earthdata_stac}
\alias{bb_handler_earthdata_stac}
\title{Handler for data sets from Earthdata providers, using the STAC catalogue to do the initial search}
\usage{
bb_handler_earthdata_stac(...)
}
\arguments{
\item{...}{: parameters, see Description}
}
\value{
A tibble with columns \code{ok}, \code{files}, \code{message}
}
\description{
This is a handler function to be used with data sets from NASA's Earthdata system. This function is not intended to be called directly, but rather is specified as a \code{method} option in \code{\link{bb_source}}.
This function finds the items to download via the STAC catalog, and then uses \code{\link{bb_rget}} to download them. Data sources using this function can provide appropriate \code{\link{bb_rget}} parameters if required.
The method arguments accepted by \code{bb_handler_aws_s3} are currently:
\itemize{
  \item "stac_id" string: the STAC identifier
  \item "collection_id" string: the collection identifier
}
}
\examples{
\dontrun{

my_source <- bb_source(
  name = "Nimbus Ice Edge Points from Nimbus Visible Imagery",
  id = "10.5067/NIMBUS/NmIcEdg2",
  description = "This data set (NmIcEdg2) ... [truncated; see sources_seaice()]",
  doc_url = "http://nsidc.org/data/nmicedg2/",
  citation = "Gallaher D and Campbell G ... [truncated; see sources_seaice()]",
  license = "Please cite, see http://nsidc.org/about/use_copyright.html",
  authentication_note = "Requires Earthdata login, see https://urs.earthdata.nasa.gov/.
    Note that you will also need to authorize the application 'nsidc-daacdata'
    (see 'My Applications' at https://urs.earthdata.nasa.gov/profile)",
  method = list("bb_handler_earthdata_stac", stac_id = "NSIDC_CPRD",
    collection_id = "NmIcEdg2_1"),
  user = "your_earthdata_username",
  password = "your_earthdata_password",
  collection_size = 0.02)
}

}
\references{
https://wiki.earthdata.nasa.gov/display/EL/How+To+Register+With+Earthdata+Login
}
