% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rget.R
\name{bb_handler_rget}
\alias{bb_handler_rget}
\title{Mirror an external data source using bowerbird's bb_rget utility}
\usage{
bb_handler_rget(...)
}
\arguments{
\item{...}{: parameters passed to \code{\link{bb_rget}}}
}
\value{
TRUE on success
}
\description{
This is a general handler function that is suitable for a range of data sets. This function is not intended to be called directly, but rather is specified as a \code{method} option in \code{\link{bb_source}}.
}
\details{
This handler function makes calls to the \code{\link{bb_rget}} function. Arguments provided to \code{bb_handler_rget} are passed through to \code{\link{bb_rget}}.
}
\examples{

my_source <- bb_source(
   id = "gshhg_coastline",
   name = "GSHHG coastline data",
   description = "A Global Self-consistent, Hierarchical, High-resolution Geography Database",
   doc_url = "http://www.soest.hawaii.edu/pwessel/gshhg",
   citation = "Wessel, P., and W. H. F. Smith, A Global Self-consistent, Hierarchical,
     High-resolution Shoreline Database, J. Geophys. Res., 101, 8741-8743, 1996",
   source_url = "ftp://ftp.soest.hawaii.edu/gshhg/*",
   license = "LGPL",
   method = list("bb_handler_rget", recursive = TRUE, level = 1,
                 accept_download = "bin.+\\\\.zip|README.TXT"),
   postprocess = list("bb_unzip"),
   collection_size = 0.6)

}
\seealso{
\code{\link{bb_rget}}, \code{\link{bb_source}}
}
