% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thredds.R
\name{bb_handler_thredds}
\alias{bb_handler_thredds}
\title{Mirror an external thredds data source using bowerbird's bb_rget utility}
\usage{
bb_handler_thredds(...)
}
\arguments{
\item{...}{: parameters passed to \code{\link{bb_rget}}}
}
\value{
TRUE on success
}
\description{
This function is not intended to be called directly, but rather is specified as a \code{method} option in \code{\link{bb_source}}.
}
\examples{
my_source <- bb_source(
    name = "OSI SAF Global Low Resolution Sea Ice Drift",
    id = "10.15770/EUM_SAF_OSI_NRT_2007",
    description = "Example dataset.",
    doc_url = "https://osi-saf.eumetsat.int/products/osi-405-c",
    source_url = "https://thredds.met.no/thredds/catalog/osisaf/met.no/ice/drift_lr/merged/2009/catalog.html",
    citation = "See https://doi.org/10.15770/EUM_SAF_OSI_NRT_2007",
    license = "Please cite",
    method = list("bb_handler_thredds", level = 2),
    access_function = "",
    data_group = "Sea ice")

\dontrun{
  result <- bb_get(my_source, local_file_root = tempdir(), verbose = TRUE)
}

}
\seealso{
\code{\link{bb_rget}}, \code{\link{bb_source}}, \code{\link{bb_sync}}
}
