% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{bb_modify_source}
\alias{bb_modify_source}
\title{Modify a data source}
\usage{
bb_modify_source(src, ...)
}
\arguments{
\item{src}{data.frame or tibble: a single-row data source (as returned by \code{bb_source})}

\item{...}{: parameters as for \code{bb_source}}
}
\value{
as for \code{bb_source}: a tibble with columns as per the \code{bb_source} function arguments (excluding \code{warn_empty_auth})
}
\description{
This is a helper function designed to make it easier to modify an already-defined data source. Generally, parameters passed here will replace existing entries in \code{src} if they exist, or will be added if not. The \code{method} and \code{postprocess} parameters are slightly different: see Details, below.
}
\details{
With the exception of the \code{method} and \code{postprocess} parameters, any parameter provided here will entirely replace its equivalent in the \code{src} object. Pass a new value of \code{NULL} to remove an existing parameter.

The \code{method} and \code{postprocess} parameters are lists, and modification for these takes place at the list-element level: any element of the new list will replace its equivalent element in the list in src. If the src list does not contain that element, it will be added. To illustrate, say that we have created a data source with:

\code{src <- bb_source(method=list("bb_handler_wget",parm1=value1,parm2=value2),...)}

Calling

\code{bb_modify_source(src,method=list(parm1=newvalue1))}

will result in a new \code{method} value of \code{list("bb_handler_wget",parm1=newvalue1,parm2=value2)}

Modifying \code{postprocess} elements is similar. Note that it is not currently possible to entirely remove a postprocess component using this function. If you need to do so, you'll need to do it manually.
}
\examples{

## this pre-defined source requires a username and password
src <- subset(bb_example_sources(),id=="10.5067/EYICLBOAAJOU")

## add username and password
src <- bb_modify_source(src,user="myusername",password="mypassword")

## or using the pipe operator
src <- subset(bb_example_sources(),id=="10.5067/EYICLBOAAJOU") \%>\%
  bb_modify_source(user="myusername",password="mypassword")

## remove the existing "data_group" component
src \%>\% bb_modify_source(data_group=NULL)

## change just the 'level' setting of an existing method definition
src \%>\% bb_modify_source(method=list(level=3))

## remove the 'level' component of an existing method definition
src \%>\% bb_modify_source(method=list(level=NULL))

}
\seealso{
\code{\link{bb_source}}
}
