% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceandata.R
\name{bb_oceandata_cleanup}
\alias{bb_oceandata_cleanup}
\title{Postprocessing: remove redundant NRT oceandata files}
\usage{
bb_oceandata_cleanup(...)
}
\arguments{
\item{...}{: extra parameters passed automatically by \code{bb_sync}}
}
\value{
a list, with components \code{status} (TRUE on success) and \code{deleted_files} (character vector of paths of files that were deleted)
}
\description{
This function is not intended to be called directly, but rather is specified as a \code{postprocess} option in \code{\link{bb_source}}.
}
\details{
This function will remove near-real-time (NRT) files from an oceandata collection that have been superseded by their non-NRT versions.
}
