% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bb_source}
\alias{bb_source}
\title{Configuration information for external data sources}
\usage{
bb_source(ids = c("seaice_smmr_ssmi_nasateam"))
}
\arguments{
\item{ids}{character vector: one or more identifiers of pre-packaged source configurations}
}
\value{
tibble
}
\description{
Configuration information for external data sources
}
\examples{
\dontrun{
  cf <- bb_config() \%>\%
    add(bb_source("seaice_smmr_ssmi_nasateam"))
}

}
\seealso{
\code{\link{bb_config}}
}
