% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{bb_source2}
\alias{bb_source2}
\title{Define a data source}
\usage{
bb_source2(id, name, description = NA_character_, reference, source_url,
  citation, license, comment = NA_character_, method, postprocess,
  authentication_note = NA_character_, user = NA_character_,
  password = NA_character_, access_function = NA_character_,
  data_group = NA_character_, collection_size = NA,
  warn_empty_auth = TRUE)
}
\arguments{
\item{id}{string: (required) a unique identifier of the data source. If the data source has a DOI, use that. Otherwise, if the original data provider has an identifier for this dataset, that is probably a good choice here (include the data version number if there is one). The ID should be something that changes when the data set changes (is updated). A DOI is ideal for this}

\item{name}{string: (required) a unique name for the data source. This should be a human-readable but still concise name}

\item{description}{string: a description of the data source}

\item{reference}{string: (required) URL to the metadata record or home page of the data source}

\item{source_url}{character vector: one or more source URLs. Required for \code{bb_handler_wget}, although some \code{method} functions might not require one}

\item{citation}{string: (required) details of the citation for the data source}

\item{license}{string: (required) description of the license. For standard licenses (e.g. creative commons) include the license descriptor ("CC-BY", etc)}

\item{comment}{string: comments about the data source. If only part of the original data collection is mirrored, mention that here}

\item{method}{list: a list object that defines the function used to synchronise this data source. The first element of the list is the function name (as a string or function). Additional list elements can be used to specify additional parameters to pass to that function. If the handler function uses wget (e.g. \code{bb_handler_wget}), these extra parameters are passed to the \code{bb_wget} function}

\item{postprocess}{list: each element of \code{postprocess} defines a postprocessing step to be run after the main synchronization has happened. Each element of this list can be a function or string function name, or a list in the style of \code{list(fun,arg1=val1,arg2=val2)} where \code{fun} is the function to be called and \code{arg1} and \code{arg2} are additional parameters to pass to that function}

\item{authentication_note}{string: if authentication is required in order to access this data source, make a note of the process (include a URL to the registration page, if possible)}

\item{user}{string: username, if required}

\item{password}{string: password, if required}

\item{access_function}{string: name of the R function that can be used to read these data}

\item{data_group}{string: the name of the group to which this data source belongs. Useful for arranging sources in terms of thematic areas}

\item{collection_size}{numeric: approximate disk space (in GB) used by the data collection, if known. If the data are supplied as compressed files, this size should reflect the disk space used after decompression. If the data_source definition contains multiple source_url entries, this size should reflect the overall disk space used by all combined}

\item{warn_empty_auth}{logical: if TRUE, issue a warning if the data source requires authentication (authentication_note is not NA) but user and password have not been provided}
}
\value{
tibble
}
\description{
Define a data source
}
\examples{

my_source <- bb_source2(
   id="gshhg_coastline",
   name="GSHHG coastline data",
   description="A Global Self-consistent, Hierarchical, High-resolution Geography Database",
   reference= "http://www.soest.hawaii.edu/pwessel/gshhg",
   citation="Wessel, P., and W. H. F. Smith, A Global Self-consistent, Hierarchical,
     High-resolution Shoreline Database, J. Geophys. Res., 101, 8741-8743, 1996",
   source_url="ftp://ftp.soest.hawaii.edu/gshhg/*",
   license="",
   comment="",
   method=list("bb_handler_wget",recursive=TRUE,level=1,accept="*bin*.zip,README.TXT"),
   postprocess=list("bb_unzip"),
   collection_size=0.6)

cf <- bb_config("/my/repo/root")
cf <- bb_add(cf,my_source)

}
\seealso{
\code{\link{bb_config}}
}
