% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_sources.R
\name{bb_source_us_buildings}
\alias{bb_source_us_buildings}
\title{Example bowerbird data source: Microsoft US Buildings}
\usage{
bb_source_us_buildings(states)
}
\arguments{
\item{states}{character: (optional) one or more US state names for which to download data. If missing, data from all states will be downloaded. See the reference page for valid state names}
}
\value{
a tibble with columns as specified by \code{\link{bb_source}}
}
\description{
This function constructs a data source definition for the Microsoft US Buildings data set. This data set contains 124,885,597 computer generated building footprints in all 50 US states. NOTE: currently, the downloaded zip files will not be unzipped automatically. Work in progress.
}
\examples{
\dontrun{
## define a configuration and add this buildings data source to it
##  only including data for the District of Columbia and Hawaii
cf <- bb_config(tempdir()) \%>\%
  bb_add(bb_source_us_buildings(states = c("District of Columbia", "Hawaii")))

## synchronize (download) the data
bb_sync(cf)
}

}
\references{
\url{https://github.com/Microsoft/USBuildingFootprints}
}
\seealso{
\code{\link{bb_example_sources}}, \code{\link{bb_config}}, \code{\link{bb_handler_wget}}
}
