% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{bb_sources}
\alias{bb_sources}
\title{Bowerbird configurations for various data sources}
\usage{
bb_sources(data_group)
}
\arguments{
\item{data_group}{character vector: only return data sources belonging to these data groups}
}
\value{
tibble
}
\description{
Bowerbird configurations for various data sources
}
\examples{
bb_sources()

}
\references{
See \code{reference} and \code{citation} field in each row of the returned tibble
}
\seealso{
\code{\link{bb_config}}
}
