% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bb_summary}
\alias{bb_summary}
\title{Produce summary of bowerbird configuration}
\usage{
bb_summary(cf, file = tempfile(fileext = ".html"), format = "html")
}
\arguments{
\item{cf}{tibble: configuration, as returned by \code{bb_config}}

\item{file}{string: path to file to write summary to. A temporary file is used by default}

\item{format}{string: produce HTML ("html") or Rmarkdown ("Rmd") file?}
}
\value{
path to the summary file in HTML or Rmarkdown format
}
\description{
Produce summary of bowerbird configuration
}
\examples{
\dontrun{
  cf <- bb_config("/my/file/root") \%>\%
    add(bb_source("seaice_smmr_ssmi_nasateam"))
  browseURL(bb_summary(cf))
}

}
