% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{bb_sync}
\alias{bb_sync}
\title{Run a bowerbird data repository synchronisation}
\usage{
bb_sync(config, create_root = FALSE, verbose = TRUE, catch_errors = TRUE)
}
\arguments{
\item{config}{bb_config: configuration as returned by \code{\link{bb_config}}}

\item{create_root}{logical: should the data root directory be created if it does not exist?}

\item{verbose}{logical: if TRUE, provide additional progress output}

\item{catch_errors}{logical: if TRUE, catch errors and continue the synchronisation process}
}
\value{
a tibble with the name, id, source_url, and sync success status of each data source (multiple source_urls within a data source will appear in separate rows)
}
\description{
For sync purposes, each row in the config's data_sources table is expanded on source_url, so that there is one row per unique data_source/source_url combination. Hence the returned tibble from \code{bb_sync} may have more rows than \code{bb_data_sources(config)}.
}
