% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bb_validate}
\alias{bb_validate}
\title{Validate a bowerbird configuration}
\usage{
bb_validate(cf)
}
\arguments{
\item{cf}{data.frame: configuration, as returned by \code{bb_config}}
}
\value{
TRUE (invisibly) or throw error
}
\description{
Runs some basic sanity checks on a bowerbird configuration.
}
\seealso{
\code{\link{bb_config}}
}
