% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wget.R
\name{bb_wget}
\alias{bb_wget}
\title{Mirror an external data source using the wget utility}
\usage{
bb_wget(data_source)
}
\arguments{
\item{data_source}{tibble: single-row tibble defining a data source, e.g. as returned by \code{bb_source}}
}
\value{
TRUE on success
}
\description{
Mirror an external data source using the wget utility
}
\seealso{
\code{\link{wget}}
}
