% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenodo.R
\name{bb_zenodo_source}
\alias{bb_zenodo_source}
\title{Generate a bowerbird data source object for a Zenodo data set}
\usage{
bb_zenodo_source(id)
}
\arguments{
\item{id}{: the ID of the data set}
}
\value{
A tibble containing the data source definition, as would be returned by \code{\link[bowerbird]{bb_source}}
}
\description{
Generate a bowerbird data source object for a Zenodo data set
}
\seealso{
\code{\link{bb_source}}
}
