% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{BOWER-class}
\alias{BOWER-class}
\alias{BOWER}
\title{BOWER}
\description{
BOWER
}
\section{Slots}{

\describe{
\item{\code{genesets}}{A list of containing vectors of genes.}

\item{\code{graph}}{An igraph object that represents the kNN graph.}

\item{\code{clusters}}{A vector holding the cluster labels for each geneset.}

\item{\code{.graph_data}}{Hidden slot for graph in dataframe format.}
}}

