% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_source_dir}
\alias{data_source_dir}
\title{Return the local directory of a data source}
\usage{
data_source_dir(data_source)
}
\arguments{
\item{data_source}{tibble: single-row tibble defining a data source, e.g. as returned by \code{bb_source}}
}
\value{
string: the directory
}
\description{
Files from that data source are stored locally in this directory.
}
\examples{
\dontrun{
  cf <- bb_config("/my/file/root") \%>\%
    add(bb_sources("NSIDC SMMR-SSM/I Nasateam sea ice concentration"))
  data_source_dir(cf)
}

}
