% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scFA.R
\name{.init_scFA}
\alias{.init_scFA}
\title{shortcut to initiate building of scFA}
\usage{
.init_scFA()
}
\value{
for internal use only. quickly redocumenting scFA
}
\description{
shortcut to initiate building of scFA
}
\examples{
\donttest{
init_scFA()
}
}
