% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_genesets.R
\name{enrich_genesets}
\alias{enrich_genesets}
\alias{enrich_genesets.list}
\alias{enrich_genesets.Seurat}
\alias{enrich_genesets.SingleCellExperiment}
\title{enrich_genesets}
\usage{
\method{enrich_genesets}{list}(
  list,
  bower,
  core = FALSE,
  gene_symbol = "X1",
  logfoldchanges = "logfoldchanges",
  pvals = "pvals",
  remove_mito_ribo = TRUE,
  minSize = 0,
  maxSize = 1000,
  ...
)

\method{enrich_genesets}{Seurat}(
  sce,
  bower,
  groupby = NULL,
  core = FALSE,
  standardize = TRUE,
  mode = c("AUCell", "Seurat", "scanpy"),
  sce_assay = "logcounts",
  seurat_assay = "RNA",
  ncpus = NULL,
  aucMaxRank_pct = 5,
  ...
)

\method{enrich_genesets}{SingleCellExperiment}(
  sce,
  bower,
  groupby = NULL,
  core = FALSE,
  standardize = TRUE,
  mode = c("AUCell", "Seurat", "scanpy"),
  sce_assay = "logcounts",
  seurat_assay = "RNA",
  ncpus = NULL,
  aucMaxRank_pct = 5,
  ...
)
}
\arguments{
\item{list}{list containing differentially expressed gene testing results in a data frame.}

\item{bower}{Processed BOWER object.}

\item{core}{boolean. Whether or not to use the coregenes of genesets slot. Default is FALSE (use genesets).}

\item{gene_symbol}{column name for gene_symbol for gsea.}

\item{logfoldchanges}{column name for logfoldchanges for gsea.}

\item{pvals}{column name for pvals for gsea.}

\item{remove_mito_ribo}{boolean. whether or not to remove mitochondial and ribosomal genes from consideration for gsea. Default is TRUE.}

\item{minSize}{minimum geneset size for gsea. Default is 0.}

\item{maxSize}{maximum geneset size for gsea. Default is 1000.}

\item{...}{passed to fgsea::fgsea, AUCell::AUCell_buildRankings or Seurat::AddModuleScore}

\item{sce}{a single cell object in the format of a Seurat object or SingleCellExperiment object.}

\item{groupby}{Column name in the meta.data/colData of the single cell objects specifying the group to average the enrichment score. If not specified, not cluster average will be calculated.}

\item{standardize}{whether or not to standardize mean enrichment values to 0 to 1. Only used if groupby is not NULL.}

\item{mode}{choice of enrichment test to perform.}

\item{sce_assay}{name of assay in SingleCellExperiment object.}

\item{seurat_assay}{name of assay in Seurat object.}

\item{ncpus}{number of cores used for parallelizing geneset testing.}

\item{aucMaxRank_pct}{percentage to use for aucMaxRank in AUCell::AUCell_calcAUC.}

\item{nperm}{number of permuation iterations for gsea. Default is 10000.}
}
\value{
Returns a dataframe of average gene set scores.
}
\description{
Runs geneset tests and store in BOWER class.
}
\details{

}
\examples{
library(ktplots)
data(kidneyimmune)
gmt_file <- system.file("extdata", "h.all.v7.4.symbols.gmt", package = "bowerbird")
bwr <- bower(gmt_file)
bwr <- snn_graph(bwr)
bwr <- find_clusters(bwr)
bwr <- summarize_clusters(bwr, ncpus = 1)
bwr <- enrich_genesets(kidneyimmune, bwr, groupby = 'celltype', ncpus = 1)
bwr
}
