% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{fame-class}
\alias{fame-class}
\alias{fame}
\title{The fame container class}
\description{
The fame container class
}
\section{Slots}{

\describe{
\item{\code{genesets}}{A list of containing vectors of genes.}

\item{\code{graph}}{An igraph object that represents the kNN graph.}
}}

