% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{find_clusters}
\alias{find_clusters}
\alias{find_clusters.BOWER}
\alias{find_clusters.igraph}
\title{clusters}
\usage{
\method{find_clusters}{BOWER}(bower, resolution = 3, ...)

\method{find_clusters}{igraph}(gr, resolution = 3, ...)
}
\arguments{
\item{bower}{BOWER class..}

\item{resolution}{value for leiden clustering.}

\item{...}{passed to leiden::leiden.}

\item{gr}{igraph object.}
}
\value{
Cluster assignment for each geneset.
}
\description{
Performs clustering of SNN graph of genesets.
}
\examples{
\donttest{
bwr <- find_clusters(bwr)
}
}
