% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wget.R
\name{install_wget}
\alias{install_wget}
\title{Helper function to install wget on Windows}
\usage{
install_wget()
}
\value{
TRUE on success
}
\description{
Helper function to install wget on Windows
}
\examples{
\dontrun{
  install_wget()
}

}
\references{
https://eternallybored.org/misc/wget/current/wget.exe
}
