% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R, R/generics.R
\name{.is_blank}
\alias{.is_blank}
\alias{.filter_geneset}
\alias{.emptyBOWER}
\alias{.create_graph}
\alias{.graph_to_data}
\alias{.closest_to_centroid}
\alias{.typecheck}
\alias{.makeRankGeneList}
\alias{range01}
\alias{.retrieve_gsea}
\alias{.read_geneset}
\title{misc}
\usage{
.is_blank(x)

.filter_geneset(gs, min_size)

.emptyBOWER(...)

.create_graph(snn, mode = "undirected", weighted = TRUE, ...)

.graph_to_data(gr, ...)

.closest_to_centroid(data, x_column, y_column, n = 1, id_column = NULL)

.typecheck(sce, mode, sce_assay = "logcounts", seurat_assay = "RNA")

.makeRankGeneList(
  deg,
  gene_symbol,
  logfoldchanges,
  pvals,
  remove_mito_ribo = TRUE
)

range01(x)

.retrieve_gsea(res, query)

.read_geneset(x, ...)
}
\arguments{
\item{x}{numerical vector}

\item{gs}{list of vectors.}

\item{min_size}{minimum size of geneset otherwise filter out.}

\item{...}{passed to ggraph::ggraph.}

\item{snn}{KNN graph.}

\item{mode}{mode of conversion. Accepts one of AUCell, SingleCellExperiment, Seurat, scanpy.}

\item{weighted}{This argument specifies whether to create a weighted graph from an adjacency matrix. If it is NULL then an unweighted graph is created and the elements of the adjacency matrix gives the number of edges between the vertices. If it is a character constant then for every non-zero matrix entry an edge is created and the value of the entry is added as an edge attribute named by the weighted argument. If it is TRUE then a weighted graph is created and the name of the edge attribute will be weight. See ?igraph::graph_from_adjacency_matrix.}

\item{gr}{data matrix or data.frame that contains at least two columns.}

\item{x_column}{name of x column/axis.}

\item{y_column}{name of y column/axis.}

\item{n}{(numeric) number of points that are close to the centroid to be detected. Default = 1.}

\item{id_column}{(character or numeric) name or numeric index of the column in data containing identifiers of one or distinct sets of points. If, NULL, the default, only one set is assumed.}

\item{sce}{Single cell object. Accepts either a Seurat object of SingleCellExperiment object.}

\item{sce_assay}{name of assay in SingleCellExperiment object.}

\item{seurat_assay}{name of assay in Seurat object.}

\item{deg}{deg table. must at least have a column for gene symbol, log fold changes and pvalues.}

\item{gene_symbol}{gene_symbol}

\item{logfoldchanges}{logfoldchanges}

\item{pvals}{pvals}

\item{remove_mito_ribo}{boolean. whether or not to remove mitochondial and ribosomal genes from consideration. Default is TRUE.}

\item{res}{fgsea result data.frame.}

\item{query}{column name in fgsea results.}
}
\value{
appropriate single cell object for enrichment analysis

sorted ranked gene list.
}
\description{
Miscellaneous utility functions.
}
