% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceandata.R
\name{oceandata_parameters}
\alias{oceandata_parameters}
\title{rdname oceandata_parameter_map}
\usage{
oceandata_parameters(platform)
}
\arguments{
\item{platform}{V for VIIRs, S for SeaWiFS, etc.}
}
\description{
rdname oceandata_parameter_map
}
