% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_graph}
\alias{plot_graph}
\alias{plot_graph.BOWER}
\alias{plot_graph.igraph}
\title{plot_graph}
\usage{
\method{plot_graph}{BOWER}(
  bower,
  colorby = "cluster",
  mode = c("default", "gsea"),
  gsea.slot = c("padj", "pval", "NES", "ES"),
  gsea.pval.cutoff = 0.25,
  node.size = "geneset_size",
  node.label = "labels",
  node.label.size = 3,
  title = NULL,
  title.size = 3,
  edge.width = "weight",
  edge.width.thickness.range = c(0, 1),
  edge.alpha = 0.25,
  dot_max = 8,
  guides.colour = FALSE
)

\method{plot_graph}{igraph}(
  gr,
  colorby = "cluster",
  node.size = "geneset_size",
  node.label = "labels",
  node.label.size = 3,
  title = NULL,
  title.size = 3,
  edge.width = "weight",
  edge.width.thickness.range = c(0, 1),
  edge.alpha = 0.25,
  dot_max = 8,
  guides.colour = FALSE,
  ...
)
}
\arguments{
\item{bower}{BOWER object}

\item{colorby}{name of attribute for colors of nodes.}

\item{node.size}{name of attribute for size of nodes or numerical value/vector.}

\item{node.label}{name of attribute for colors of nodes.}

\item{node.label.size}{size of node label.}

\item{edge.width}{name of attribute for width of edges.}

\item{edge.width.thickness.range}{scale for range of widths of edges.}

\item{edge.alpha}{transparency of edges.}

\item{guides.colour}{boolean. whether or not to display guide for vertex colour.}

\item{...}{passed to ggraph::ggraph}

\item{node.max_size}{maximum node size.}
}
\value{
Returns a ggraph plot.
}
\description{
Plot the geneset kNN network as a graph.
}
\details{
Just a ggraph wrapper to plot the graph.
}
\examples{
gmt_file <- system.file("extdata", "h.all.v7.4.symbols.gmt", package = "bowerbird")
bwr <- bower(gmt_file)
bwr <- snn_graph(bwr)
bwr <- find_clusters(bwr)
bwr <- summarize_clusters(bwr, ncpus = 1)
plot_graph(bwr)
}
