% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{pp_cleanup}
\alias{pp_cleanup}
\title{Postprocessing: remove unwanted files}
\usage{
pp_cleanup(data_source, pattern, recursive = FALSE, ignore_case = FALSE,
  ...)
}
\arguments{
\item{data_source}{data.frame: single-row data.frame defining a data source, e.g. as returned by \code{bb_source}}

\item{pattern}{string: regular expression, passed to \code{file.info}}

\item{recursive}{logical: should the cleanup recurse into subdirectories?}

\item{ignore_case}{logical: should pattern matching be case-insensitive?}

\item{...}{: additional arguments (currently ignored)}
}
\value{
TRUE on success
}
\description{
Postprocessing: remove unwanted files
}
\examples{
\dontrun{
  ## remove .asc files after synchronisation
  my_source <- bb_source(...,postprocess=quote(pp_cleanup,pattern="\\\\.asc$"))
}

}
\seealso{
\code{\link{bb_source}} \code{\link{bb_config}} \code{\link{pp_decompress}}
}
