% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{pp_decompress}
\alias{pp_bunzip2}
\alias{pp_decompress}
\alias{pp_gunzip}
\alias{pp_uncompress}
\alias{pp_unzip}
\title{Postprocessing: decompress zip, gz, bz2, Z files and optionally delete the compressed copy
\code{pp_unzip}, \code{pp_*} are convenience wrappers around \code{pp_decompress} that specify the method}
\usage{
pp_decompress(data_source, delete = FALSE, file_list_before, file_list_after,
  method)

pp_unzip(...)

pp_gunzip(...)

pp_bunzip2(...)

pp_uncompress(...)
}
\arguments{
\item{data_source}{tibble: single-row tibble defining a data source, e.g. as returned by \code{bb_source}}

\item{delete}{logical: delete the zip files after extracting their contents?}

\item{file_list_before}{data.frame: files present in the directory before synchronising, as returned by \code{file.info}. Only required if delete=TRUE}

\item{file_list_after}{data.frame: files present in the directory after synchronising, as returned by \code{file.info}. Only required if delete=TRUE}

\item{method}{string: one of "unzip","gunzip","bunzip2","decompress"}

\item{...}{: arguments passed to \code{pp_decompress}}
}
\value{
TRUE on success
}
\description{
Postprocessing: decompress zip, gz, bz2, Z files and optionally delete the compressed copy
\code{pp_unzip}, \code{pp_*} are convenience wrappers around \code{pp_decompress} that specify the method
}
\examples{

}
\seealso{
\code{\link{bb_source}} \code{\link{bb_config}} \code{\link{pp_cleanup}}
}

