% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapid_reponse_handler.R
\name{rapid_response_get}
\alias{rapid_response_get}
\title{Handler for NASA MODIS Rapid Response files}
\usage{
rapid_response_get(data_source)
}
\arguments{
\item{data_source}{tibble: single-row tibble defining a data source, e.g. as returned by \code{bb_source}}
}
\value{
TRUE on success
}
\description{
Handler for NASA MODIS Rapid Response files
}
\references{
https://earthdata.nasa.gov/earth-observation-data/near-real-time/rapid-response
}
