% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geneset.R
\name{read_geneset}
\alias{read_geneset}
\title{read_geneset}
\usage{
read_geneset(x, min_size = 15)
}
\arguments{
\item{x}{file,}

\item{min_size}{minimum size of geneset otherwise filter out.}
}
\value{
Returns a list of gene sets
}
\description{
read_geneset
}
\examples{
\donttest{
gs <- read_geneset('geneset.gmt')
}
}
