% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_scFA.R, R/scFA.R
\name{scFA}
\alias{scFA}
\alias{init_scFA}
\title{documenting scFA shortcut}
\usage{
init_scFA()

scFA()
}
\value{
for internal use only. quickly redocumenting scFA

for internal use only. runs init scFA
}
\description{
documenting scFA shortcut

shortcut to initiate building of scFA
}
\examples{
\donttest{
init_scFA() # leaves folder, install, and change back to folder
}
\donttest{
scFA()
}
}
