% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{set_clusters}
\alias{set_clusters}
\alias{set_clusters.BOWER}
\title{Set cluster to BOWER class}
\usage{
\method{set_clusters}{BOWER}(bower, clusters)
}
\arguments{
\item{bower}{bower class}

\item{clusters}{Cluster assignment for each geneset.}
}
\value{
bower class
}
\description{
Manually sets externally/alternatively determined cluster assignment for each geneset.
}
\examples{
\donttest{
bower <- set_clusters(bower, c(1,1,1,2,2,2,3,3,3))
}
}
