% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{slice.bb_cf}
\alias{slice.bb_cf}
\alias{[.bb_cf}
\title{Select a subset of data sources within a bowerbird configuration}
\usage{
\method{slice}{bb_cf}(.data, ...)

\method{[}{bb_cf}(x, i, j, ...)
}
\arguments{
\item{.data}{list: configuration, as returned by \code{bb_config}}

\item{...}{: integer row values}

\item{x}{list: configuration, as returned by \code{bb_config}}

\item{i}{numeric: integer row values}

\item{j}{numeric: ignored}
}
\value{
configuration
}
\description{
Select a subset of data sources within a bowerbird configuration
}
\examples{
\dontrun{
  cf <- bb_config("/my/file/root") \%>\%
    add(bb_sources()) \%>\%
    slice(1:5)
}

}
