% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{snn_graph}
\alias{snn_graph}
\alias{snn_graph.BOWER}
\alias{snn_graph.list}
\title{snn_graph}
\usage{
\method{snn_graph}{BOWER}(bower, max_features = 100, remove_stopwords = FALSE, k = 5, ...)

\method{snn_graph}{list}(gs, max_features = 100, remove_stopwords = FALSE, k = 5, ...)
}
\arguments{
\item{max_features}{use top features sorted by count to be used in bag of words matrix. The default value is set to 100.}

\item{remove_stopwords}{a list of stopwords to use, by default it uses its inbuilt list of standard stopwords. The default value is FALSE.}

\item{k}{the maximum number of nearest neighbors to search. The default value is set to 5.}

\item{...}{passed to superml::TfIdfVectorizer}

\item{gs}{genesets in list.}
}
\value{
Returns a matrix of tf-idf score of tokens.
}
\description{
Creates a k-nearest-neighbor graph from a tf-idf matrix.
}
\details{
Given a list of text, it creates a sparse matrix consisting of tf-idf score for tokens from the text. See \verb{https://github.com/saraswatmks/superml/blob/master/R/TfidfVectorizer.R}. A k shortest-nearest neighbor graph is then computed using the overlap of of the terms.
}
\examples{
gmt_file <- system.file("extdata", "h.all.v7.4.symbols.gmt", package = "bowerbird")
bwr <- bower(gmt_file)
bwr <- snn_graph(bwr)
}
