% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_clusters.R
\name{summarize_clusters.BOWER}
\alias{summarize_clusters.BOWER}
\alias{summarize_clusters.igraph}
\title{summarize_clusters}
\usage{
\method{summarize_clusters}{BOWER}(
  bower,
  cluster = NULL,
  pattern = NULL,
  sep = NULL,
  ncpus = NULL,
  disconnect_graph = FALSE,
  ...
)

\method{summarize_clusters}{igraph}(
  graph,
  cluster = NULL,
  pattern = NULL,
  sep = NULL,
  ncpus = NULL,
  disconnect_graph = FALSE,
  ...
)
}
\arguments{
\item{cluster}{vector of cluster labels for each geneset.}

\item{pattern}{search pattern to remove from the terms. Unless specified, will default to built-in pattern.}

\item{sep}{separator used/found in gene set names to be changed to blank spaces. Default value is underscore ('_').}

\item{ncpus}{number of cores used for parallelizing reconstruction.}

\item{disconnect_graph}{return a graph connecting only nodes in a cluster.}

\item{...}{passed to textrank::textrank_sentences.}

\item{graph}{geneset overlap graph.}
}
\value{
Returns a matrix of tf-idf score of tokens.
}
\description{
Summarize the terms of the cluster using pagerank algorithm
}
\details{
Given a list of text, it creates a sparse matrix consisting of tf-idf score for tokens from the text. See \verb{https://github.com/saraswatmks/superml/blob/master/R/TfidfVectorizer.R}. A k shortest-nearest neighbor graph is then computed using the overlap of of the terms.
}
\examples{
\donttest{
terms <- summarize_clusters(mat)
}
}
