% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadc_eds_handler.R
\name{aadc_eds_get}
\alias{aadc_eds_get}
\title{Handler for files downloaded from the Australian Antarctic Data Centre EDS system}
\usage{
aadc_eds_get(data_source)
}
\arguments{
\item{data_source}{tibble: single-row tibble defining a data source, e.g. as returned by \code{bb_source}}
}
\value{
TRUE on success
}
\description{
Handler for files downloaded from the Australian Antarctic Data Centre EDS system
}
\references{
http://data.aad.gov.au
}
