% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aadc_eds_handler.R
\name{aadc_eds_get}
\alias{aadc_eds_get}
\title{Handler for files downloaded from the Australian Antarctic Data Centre EDS system}
\usage{
aadc_eds_get(data_source, verbose = FALSE, local_dir_only = FALSE)
}
\arguments{
\item{data_source}{data.frame: single-row data.frame defining a data source, e.g. as returned by \code{bb_source}}

\item{verbose}{logical: if TRUE, provide additional progress output}

\item{local_dir_only}{logical: if TRUE, just return the local directory into which files from this data source would be saved}
}
\value{
TRUE on success
}
\description{
AADC EDS files have a URL of the form https://data.aad.gov.au/eds/file/wxyz/ or https://data.aad.gov.au/eds/wxyz/download where wxyz is a numeric file identifier.
}
\references{
http://data.aad.gov.au
}
