% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{add}
\alias{add}
\title{Add a new data source to a bowerbird configuration}
\usage{
add(cf, source)
}
\arguments{
\item{cf}{tibble: configuration, as returned by \code{bb_config}}

\item{source}{tibble: data source definition to add to the configuration, as returned by \code{bb_source}}
}
\value{
configuration tibble
}
\description{
Add a new data source to a bowerbird configuration
}
\examples{
\dontrun{
  cf <- bb_config("/my/file/root") \%>\%
    add(bb_source("seaice_smmr_ssmi_nasateam"))
}
}
\seealso{
\code{\link{bb_source}} \code{\link{bb_config}}
}

