% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{add}
\alias{add}
\title{Add a new data source to a bowerbird configuration}
\usage{
add(cf, source)
}
\arguments{
\item{cf}{tibble: configuration, as returned by \code{bb_config}}

\item{source}{tibble: data source definition to add to the configuration, as returned by \code{bb_source}}
}
\value{
configuration tibble
}
\description{
To do: include details of how this is done, and how global options are treated.
}
\seealso{
\code{\link{bb_source}} \code{\link{bb_config}}
}
