% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bb_attributes}
\alias{bb_attributes}
\title{Returns a bowerbird configuration object's settings}
\usage{
bb_attributes(config)
}
\arguments{
\item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config})}
}
\value{
named list
}
\description{
These are repository-wide settings that are applied to all data sources added to the configuration.
}
\examples{
cf <- bb_config(local_file_root="/your/data/directory")
bb_attributes(cf)

}
\seealso{
\code{\link{bb_config}}
}
