% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bb_config}
\alias{bb_config}
\title{Initialize a bowerbird configuration}
\usage{
bb_config(local_file_root, wget_default_flags = NULL,
  wget_global_flags = "--restrict-file-names=windows --progress=dot:giga",
  http_proxy = NULL, ftp_proxy = NULL, clobber = 1)
}
\arguments{
\item{local_file_root}{string: location of data repository on local file system}

\item{wget_default_flags}{string: default flags to be passed to wget. These are overridden on a per-data source basis if the data source defines its own wget_flags}

\item{wget_global_flags}{string: wget flags that will be applied to all data sources. These will be appended to either the data source wget flags (if specified), or the wget_default_flags}

\item{http_proxy}{string: URL of HTTP proxy to use e.g. 'http://your.proxy:8080' (NULL for no proxy)}

\item{ftp_proxy}{string: URL of FTP proxy to use e.g. 'http://your.proxy:21' (NULL for no proxy)}

\item{clobber}{numeric: 0=do not overwrite existing files, 1=overwrite if the remote file is newer than the local copy, 2=always overwrite existing files. For data sources that use method 'wget', an appropriate flag will be added to the wget call according to the clobber setting ("--no-clobber" to not overwrite existing files, "--timestamping" to overwrite if the remote file is newer than the local copy)}
}
\value{
configuration tibble
}
\description{
Initialize a bowerbird configuration
}
\examples{
\dontrun{
  cf <- bb_config("/my/file/root") \%>\%
    add(bb_source("seaice_smmr_ssmi_nasateam"))

  ## save to file
  saveRDS(cf,file="my_config.rds")
  ## load previously saved config
  cf <- readRDS(file="my_config.rds")
}

}
\seealso{
\code{\link{bb_source}}
}
