% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bb_config}
\alias{bb_config}
\alias{bb_save_config}
\title{Load or initialize bowerbird configuration}
\usage{
bb_config(local_file_root, wget_flags = "--progress=dot:giga",
  http_proxy = NULL, ftp_proxy = NULL, clobber = 1, file)

bb_save_config(cf, file)
}
\arguments{
\item{local_file_root}{string: location of data repository on local file system}

\item{wget_flags}{string: default flags to be passed to wget. These are overridden on a per-data source basis if the data source defines its own wget_flags}

\item{http_proxy}{string: URL of HTTP proxy to use e.g. 'http://your.proxy:8080' (NULL for no proxy)}

\item{ftp_proxy}{string: URL of FTP proxy to use e.g. 'http://your.proxy:21' (NULL for no proxy)}

\item{clobber}{numeric: 0=do not overwrite existing files, 1=overwrite if the remote file is newer than the local copy, 2=always overwrite existing files. For data sources that use method 'wget', an appropriate flag will be added to the wget call according to the clobber setting ("--no-clobber" to not overwrite existing files, "--timestamping" to overwrite if the remote file is newer than the local copy)}

\item{file}{string: file name of configuration to load (for \code{bb_config}) or save to (for \code{bb_save_config}). For \code{bb_config}, all other parameters are ignored if \code{file} is provided}

\item{cf}{tibble: configuration, as returned by \code{bb_config}}
}
\value{
configuration tibble
}
\description{
If \code{file} is provided to \code{bb_config}, load that configuration. Otherwise initialize an empty configuration.
}
\examples{
\dontrun{
  cf <- bb_config("/my/file/root") \%>\%
    add(bb_source("seaice_smmr_ssmi_nasateam"))

  ## save to file
  bb_save_config(cf,file="my_config.rds")
  ## load previously saved config
  cf <- bb_config(file="my_config.rds")
}

}
\seealso{
\code{\link{bb_source}}
}
