% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bb_data_sources}
\alias{bb_data_sources}
\alias{bb_data_sources<-}
\title{Gets or sets a bowerbird configuration object's data sources}
\usage{
bb_data_sources(config)

bb_data_sources(config) <- value
}
\arguments{
\item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config})}

\item{value}{data.frame: new data sources to set (e.g. as returned by \code{bb_example_sources}}
}
\value{
data.frame
}
\description{
Gets or sets a bowerbird configuration object's data sources
}
\examples{
cf <- bb_config(local_file_root="/your/data/directory")
cf <- bb_add(cf,bb_example_sources())
bb_data_sources(cf)

}
\seealso{
\code{\link{bb_config}}
}
