% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_sources.R
\name{bb_example_sources}
\alias{bb_example_sources}
\title{Example bowerbird data sources}
\usage{
bb_example_sources(sources)
}
\arguments{
\item{sources}{character: names or identifiers of one or more sources to return. See Details for the list of example sources and a brief explanation of each}
}
\value{
a tibble with columns as specified by \code{\link{bb_source}}
}
\description{
These example sources are useful as data sources in their own right, but are primarily provided as demonstrations of how to define data sources. See also \code{vignette("bowerbird")} for further examples and discussion.
}
\details{
Example data sources:
\itemize{
  \item "NOAA OI SST V2" - a straightforward data source that requires a simple one-level recursive download
  \item "Australian Election 2016 House of Representatives data" - an example of a recursive download that uses additional criteria to restrict what is downloaded
  \item "CMEMS global gridded SSH reprocessed (1993-ongoing)" - a data source that requires a username and password
  \item "Oceandata SeaWiFS Level-3 mapped monthly 9km chl-a" - an example data source that uses the \code{bb_handler_oceandata} method
  \item "Sea Ice Trends and Climatologies from SMMR and SSM/I-SSMIS, Version 3" - an example data source that uses the \code{bb_handler_earthdata} method
  \item "Bathymetry of Lake Superior" - another example that passes extra flags to the \code{bb_handler_rget} call in order to restrict what is downloaded
}
}
\examples{
## define a configuration and add the 2016 election data source to it
cf <- bb_config("/my/file/root") \%>\% bb_add(
   bb_example_sources("Australian Election 2016 House of Representatives data"))

\dontrun{
  ## synchronize (download) the data
  bb_sync(cf)
}
}
\references{
See the \code{doc_url} and \code{citation} field in each row of the returned tibble for references associated with these particular data sources
}
\seealso{
\code{\link{bb_config}}, \code{\link{bb_handler_rget}}, \code{\link{bb_handler_oceandata}}, \code{\link{bb_handler_earthdata}}, \code{\link{bb_source_us_buildings}}
}
