% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{bb_fingerprint}
\alias{bb_fingerprint}
\title{Fingerprint the files associated with a data source}
\usage{
bb_fingerprint(config, hash = "sha1", verbose = TRUE)
}
\arguments{
\item{config}{data.frame: configuration as returned by \code{\link{bb_config}}}

\item{hash}{string: algorithm to use to calculate file hashes: "md5", "sha1", or "none". Note that file hashing can be slow for large file collections}

\item{verbose}{logical: if TRUE, provide additional progress output}
}
\value{
data.frame
}
\description{
Fingerprint the files associated with a data source
}
\examples{
\dontrun{
  cf <- bb_config("/my/file/root") \%>\%
    bb_add(bb_example_sources())
  bb_fingerprint(cf)
}

}
