% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus_handler.R
\name{bb_handler_copernicus}
\alias{bb_handler_copernicus}
\title{Handler for Copernicus Marine datasets}
\usage{
bb_handler_copernicus(product, ctype, ...)
}
\arguments{
\item{product}{string: the desired Copernicus marine product. See \code{\link[CopernicusMarine]{cms_products_list}}}

\item{ctype}{legacy parameter, ignored}

\item{...}{: parameters passed to \code{\link{bb_rget}}}
}
\value{
TRUE on success
}
\description{
This is a handler function to be used with data sets from Copernicus Marine. This function is not intended to be called directly, but rather is specified as a \code{method} option in \code{\link{bb_source}}.
}
\details{
Note that users will need a Copernicus login.
}
\references{
https://help.marine.copernicus.eu/en/collections/4060068-copernicus-marine-toolbox
}
