% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthdata_handler.R
\name{bb_handler_earthdata}
\alias{bb_handler_earthdata}
\title{Handler for data sets from Earthdata providers}
\usage{
bb_handler_earthdata(...)
}
\arguments{
\item{...}{: parameters passed to \code{\link{bb_wget}}}
}
\value{
TRUE on success
}
\description{
This is a handler function to be used with data sets from NASA's Earthdata system. This function is not intended to be called directly, but rather is specified as a \code{postprocess} option in \code{\link{bb_source}}.
}
\details{
This function uses \code{\link{bb_wget}}, and so data sources using this function will need to provide appropriate \code{\link{bb_wget}} parameters.
}
\examples{
\dontrun{

## note that the full version of this data source is provided as part of bb_example_data_sources()

my_source <- bb_source(
  name="Sea Ice Trends and Climatologies from SMMR and SSM/I-SSMIS, Version 2",
  id="10.5067/EYICLBOAAJOU",
  description="NSIDC provides this data set ... [truncated; see bb_example_data_sources()]",
  doc_url="https://nsidc.org/data/NSIDC-0192/versions/2",
  citation="Stroeve, J. and W. Meier. 2017. ... [truncated; see bb_example_data_sources()]",
  source_url=c("https://daacdata.apps.nsidc.org/pub/DATASETS/nsidc0192_seaice_trends_climo_v2/"),
  license="Please cite, see http://nsidc.org/about/use_copyright.html",
  authentication_note="Requires Earthdata login, see https://urs.earthdata.nasa.gov/.
    Note that you will also need to authorize the application 'nsidc-daacdata'
    (see 'My Applications' at https://urs.earthdata.nasa.gov/profile)",
  method=list("bb_handler_earthdata",recursive=TRUE,level=4,no_parent=TRUE,relative=TRUE),
  user="your_earthdata_username",
  password="your_earthdata_password",
  collection_size=0.02)
}

}
\references{
https://wiki.earthdata.nasa.gov/display/EL/How+To+Register+With+Earthdata+Login
}
