% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthdata_handler.R
\name{bb_handler_earthdata}
\alias{bb_handler_earthdata}
\title{Handler for files files served up through Earthdata providers}
\usage{
bb_handler_earthdata(config, verbose = FALSE, local_dir_only = FALSE)
}
\arguments{
\item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source}

\item{verbose}{logical: if TRUE, provide additional progress output}

\item{local_dir_only}{logical: if TRUE, just return the local directory into which files from this data source would be saved}
}
\value{
the directory if local_dir_only is TRUE, otherwise TRUE on success
}
\description{
Handler for files files served up through Earthdata providers
}
\references{
https://wiki.earthdata.nasa.gov/display/EL/How+To+Register+With+Earthdata+Login
}
