% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rget.R
\name{bb_handler_rget}
\alias{bb_handler_rget}
\title{Mirror an external data source using bowerbird's bb_rget utility}
\usage{
bb_handler_rget(...)
}
\arguments{
\item{...}{: parameters passed to \code{\link{bb_rget}}}
}
\value{
TRUE on success
}
\description{
This is a general handler function that is suitable for a range of data sets. This function is not intended to be called directly, but rather is specified as a \code{method} option in \code{\link{bb_source}}.
}
\details{
This handler function makes calls to the \code{\link{bb_rget}} function. Arguments provided to \code{bb_handler_rget} are passed through to \code{\link{bb_rget}}.
}
\examples{
my_source <- bb_source(
   name = "Australian Election 2016 House of Representatives data",
   id = "aus-election-house-2016",
   description = "House of Representatives results from the 2016 Australian election.",
   doc_url = "http://results.aec.gov.au/",
   citation = "Copyright Commonwealth of Australia 2017. As far as practicable, material for
               which the copyright is owned by a third party will be clearly labelled. The
               AEC has made all reasonable efforts to ensure that this material has been
               reproduced on this website with the full consent of the copyright owners.",
   source_url = "http://results.aec.gov.au/20499/Website/HouseDownloadsMenu-20499-Csv.htm",
   license = "CC-BY",
   method = list("bb_handler_rget", level = 1, accept_download = "csv$"),
   collection_size = 0.01)

my_data_dir <- tempdir()
cf <- bb_config(my_data_dir)
cf <- bb_add(cf, my_source)

\dontrun{
bb_sync(cf, verbose = TRUE)
}

}
\seealso{
\code{\link{bb_rget}}, \code{\link{bb_source}}, \code{\link{bb_sync}}
}
