% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wget.R
\name{bb_handler_wget}
\alias{bb_handler_wget}
\title{Mirror an external data source using the wget utility}
\usage{
bb_handler_wget(config, verbose = FALSE, local_dir_only = FALSE, ...)
}
\arguments{
\item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source}

\item{verbose}{logical: if TRUE, provide additional progress output}

\item{local_dir_only}{logical: if TRUE, just return the local directory into which files from this data source would be saved}

\item{...}{: extra arguments passed through to \code{bb_wget}}
}
\value{
the directory if local_dir_only is TRUE, otherwise TRUE on success
}
\description{
This function is not intended to be called directly by the user, but rather will be called internally by the \code{bb_sync} function. The typical usage of \code{bb_handler_wget} is to specify it in the \code{method} parameter of a definition: see the example below.
}
\examples{

my_source <- bb_source(
   id="gshhg_coastline",
   name="GSHHG coastline data",
   description="A Global Self-consistent, Hierarchical, High-resolution Geography Database",
   doc_url= "http://www.soest.hawaii.edu/pwessel/gshhg",
   citation="Wessel, P., and W. H. F. Smith, A Global Self-consistent, Hierarchical,
     High-resolution Shoreline Database, J. Geophys. Res., 101, 8741-8743, 1996",
   source_url="ftp://ftp.soest.hawaii.edu/gshhg/*",
   license="LGPL",
   method=list("bb_handler_wget",recursive=TRUE,level=1,accept="*bin*.zip,README.TXT"),
   postprocess=list("bb_unzip"),
   collection_size=0.6)

}
\seealso{
\code{\link{bb_wget}}, \code{\link{bb_source}}
}
