% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wget.R
\name{bb_install_wget}
\alias{bb_install_wget}
\title{Helper function to install wget}
\usage{
bb_install_wget()
}
\value{
the path to the installed executable
}
\description{
Currently only works on Windows platforms. The wget.exe executable will be downloaded from https://eternallybored.org/misc/wget/current/wget.exe and installed into your appdata directory (typically something like C:/Users/username/AppData/Roaming/)
}
\examples{
\dontrun{
  bb_install_wget()
}

}
\references{
https://eternallybored.org/misc/wget/current/wget.exe
}
\seealso{
\code{\link{bb_find_wget}}
}
