% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bb_settings}
\alias{bb_settings}
\alias{bb_settings<-}
\title{Gets or sets a bowerbird configuration object's settings}
\usage{
bb_settings(config)

bb_settings(config) <- value
}
\arguments{
\item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config})}

\item{value}{list: new values to set}
}
\value{
named list
}
\description{
These are repository-wide settings that are applied to all data sources added to the configuration.
}
\examples{
cf <- bb_config(local_file_root="/your/data/directory")

## see current settings
bb_settings(cf)

## add an http proxy
bb_settings(cf) <- list(http_proxy="http://my.proxy")

## change the current local_file_root setting
bb_settings(cf) <- list(local_file_root="/new/location")

}
\seealso{
\code{\link{bb_config}}
}
