% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{bb_source}
\alias{bb_source}
\title{Configuration information for external data sources}
\usage{
bb_source(id, name, description, reference, source_urls, citation, license,
  comment, method, method_flags, postprocess, access_function, data_group)
}
\arguments{
\item{id}{string: a short, unique identifier for the data source}

\item{name}{string:}

\item{description}{string:}

\item{reference}{string:}

\item{source_urls}{character vector:}

\item{citation}{string:}

\item{license}{string:}

\item{comment}{string:}

\item{method}{string:}

\item{method_flags}{string:}

\item{postprocess}{string or function:}

\item{access_function}{string:}

\item{data_group}{string:}
}
\value{
tibble
}
\description{
Configuration information for external data sources
}
\examples{
\dontrun{
  cf <- bb_config() \%>\%
    add(bb_sources() \%>\% dplyr::filter(id=="seaice_smmr_ssmi_nasateam"))
}

}
\seealso{
\code{\link{bb_config}}
}
