% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{bb_source}
\alias{bb_source}
\title{Define an external data source}
\usage{
bb_source(name, description = "", reference, source_urls, citation, license,
  comment = "", method = "wget", method_flags = "", postprocess,
  access_function = "", data_group = "")
}
\arguments{
\item{name}{string: the name of the data source}

\item{description}{string: a description of the data source}

\item{reference}{string: URL to the metadata record or home page of the data source}

\item{source_urls}{character vector: one or more source URLs}

\item{citation}{string:}

\item{license}{string:}

\item{comment}{string:}

\item{method}{string:}

\item{method_flags}{string:}

\item{postprocess}{function, call, or list thereof: functions to apply after synchronisation has completed. If NULL or an empty list, no postprocessing will be applied}

\item{access_function}{string:}

\item{data_group}{string:}
}
\value{
tibble
}
\description{
Define an external data source
}
\examples{

my_source <- bb_source(
   name="GSHHG coastline data",
   description="A Global Self-consistent, Hierarchical, High-resolution Geography Database",
   reference= "http://www.soest.hawaii.edu/pwessel/gshhg",
   citation="Wessel, P., and W. H. F. Smith, A Global Self-consistent, Hierarchical,
     High-resolution Shoreline Database, J. Geophys. Res., 101, 8741-8743, 1996",
   source_urls="ftp://ftp.soest.hawaii.edu/gshhg/*",
   license="",
   comment="",
   method="wget",
   method_flags="--recursive --level=1 --accept=\\"*bin*.zip,README.TXT\\"",
   postprocess=pp_unzip)

cf <- bb_config("/my/repo/root")
cf <- add(cf,my_source)

}
\seealso{
\code{\link{bb_config}}
}
