% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{bb_source}
\alias{bb_source}
\title{Define an external data source}
\usage{
bb_source(name, description = NA_character_, reference, source_url, citation,
  license, comment = NA_character_, method = bb_wget,
  method_flags = NA_character_, postprocess,
  authentication_note = NA_character_, user = NA_character_,
  password = NA_character_, access_function = NA_character_,
  data_group = NA_character_, collection_size = NA,
  warn_empty_auth = TRUE)
}
\arguments{
\item{name}{string: the name of the data source}

\item{description}{string: a description of the data source}

\item{reference}{string: URL to the metadata record or home page of the data source}

\item{source_url}{character vector: one or more source URLs}

\item{citation}{string: details of the citation for the data source}

\item{license}{string: description of the license. For standard licenses (e.g. creative commons) include the license descriptor ("CC-BY", etc)}

\item{comment}{string: comments about the data source. If only part of the original data collection is mirrored, mention that here}

\item{method}{function, call, or symbol: the function that handles the synchronisation process for this data source}

\item{method_flags}{string: flags to pass to the method. If method is \code{bb_wget}, these are wget flags. Run \code{wget("--help")} to get help on these flags}

\item{postprocess}{function, call, symbol, or list thereof: functions to apply after synchronisation has completed. If NULL or an empty list, no postprocessing will be applied}

\item{authentication_note}{string: if authentication is required in order to access this data source, make a note of the process (include a URL to the registration page, if possible)}

\item{user}{string: username, if required}

\item{password}{string: password, if required}

\item{access_function}{string: name of the R function that can be used to read these data}

\item{data_group}{string: the name of the group to which this data source belongs. Useful for arranging sources in terms of thematic areas}

\item{collection_size}{numeric: approximate disk space (in GB) used by the data collection, if known. If the data are supplied as compressed files, this size should reflect the disk space used after decompression}

\item{warn_empty_auth}{logical: if TRUE, issue a warning if the data source requires authentication (authentication_note is not NA) but user and password have not been provided}
}
\value{
tibble
}
\description{
Define an external data source
}
\examples{

my_source <- bb_source(
   name="GSHHG coastline data",
   description="A Global Self-consistent, Hierarchical, High-resolution Geography Database",
   reference= "http://www.soest.hawaii.edu/pwessel/gshhg",
   citation="Wessel, P., and W. H. F. Smith, A Global Self-consistent, Hierarchical,
     High-resolution Shoreline Database, J. Geophys. Res., 101, 8741-8743, 1996",
   source_url="ftp://ftp.soest.hawaii.edu/gshhg/*",
   license="",
   comment="",
   method=quote(bb_wget),
   method_flags="--recursive --level=1 --accept=\\"*bin*.zip,README.TXT\\"",
   postprocess=quote(pp_unzip),
   collection_size=620)

cf <- bb_config("/my/repo/root")
cf <- add(cf,my_source)

}
\seealso{
\code{\link{bb_config}}
}
