% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{bb_summary}
\alias{bb_summary}
\title{Produce a summary of a bowerbird configuration}
\usage{
bb_summary(
  config,
  file = tempfile(fileext = ".html"),
  format = "html",
  inc_license = TRUE,
  inc_auth = TRUE,
  inc_size = TRUE,
  inc_access_function = TRUE,
  inc_path = TRUE
)
}
\arguments{
\item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config})}

\item{file}{string: path to file to write summary to. A temporary file is used by default}

\item{format}{string: produce HTML ("html") or Rmarkdown ("Rmd") file?}

\item{inc_license}{logical: include each source's license and citation details?}

\item{inc_auth}{logical: include information about authentication for each data source (if applicable)?}

\item{inc_size}{logical: include each source's size (disk space) information?}

\item{inc_access_function}{logical: include each source's access function?}

\item{inc_path}{logical: include each source's local file path?}
}
\value{
path to the summary file in HTML or Rmarkdown format
}
\description{
This function produces a summary of a bowerbird configuation in HTML or Rmarkdown format. If you are maintaining a data collection on behalf of other users, or even just for yourself, it may be useful to keep an up-to-date HTML summary of your repository in an accessible location. Users can refer to this summary to see which data are in the repository and some details about them.
}
\examples{
\dontrun{
  cf <- bb_config("/my/file/root") \%>\%
    bb_add(bb_example_sources())
  browseURL(bb_summary(cf))
}

}
