% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{bb_sync}
\alias{bb_sync}
\title{Run a bowerbird data repository synchronisation}
\usage{
bb_sync(config, create_root = FALSE, verbose = TRUE)
}
\arguments{
\item{config}{tibble: configuration as returned by \code{\link{bb_config}}}

\item{create_root}{logical: should the data root directory be created if it does not exist?}

\item{verbose}{logical: if TRUE, provide additional progress output}
}
\value{
vector of logical values indicating success of each data source in config
}
\description{
Run a bowerbird data repository synchronisation
}
