% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wget.R
\name{bb_wget}
\alias{bb_wget}
\title{Make a wget call}
\usage{
bb_wget(url, flags = character(), verbose = FALSE, capture_stdout = FALSE)
}
\arguments{
\item{url}{string: the URL to retrieve}

\item{flags}{character: character vector of command-line flags to pass to wget}

\item{verbose}{logical: print trace output?}

\item{capture_stdout}{logical: if TRUE, return 'stdout' and 'stderr' output in the returned object (see exec_internal from the sys package). Otherwise send these outputs to the console}
}
\value{
the result of the system call (or if bb_wget("--help") was called, a message will be issued). The returned object will have components 'status' and (if capture_stdout was TRUE) 'stdout' and 'stderr'
}
\description{
The wget system call is made using the \code{exec_wait} function from the sys package. Call \code{bb_wget("--help")} to get a message giving information about wget's command line parameters
}
\examples{
\dontrun{
  ## get help about wget command line parameters
  bb_wget("--help")
}

}
\seealso{
\code{\link{bb_install_wget}} \code{\link{bb_find_wget}}
}
