% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bower.R
\name{bower}
\alias{bower}
\title{bower}
\usage{
bower(genesets, graph = NULL, clusters = NULL, ...)
}
\description{
A flexible function to initiate a BOWER class from a geneset file or a list of genesets.
}
\examples{
gmt_file <- system.file("extdata", "h.all.v7.4.symbols.gmt", package = "bowerbird")
bwr <- bower(gmt_file)
bwr
}
