% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{genesets-class}
\alias{genesets-class}
\alias{clusters-class}
\alias{graph-class}
\alias{coregenes-class}
\alias{scores-class}
\alias{BOWER-class}
\alias{BOWER}
\title{BOWER}
\description{
The BOWER class for Functional annotation + Gene Module Summarization
}
\details{
The BOWER container class
}
\section{Slots}{

\describe{
\item{\code{genesets}}{A list of containing vectors of genes.}

\item{\code{graph}}{An igraph object that represents the kNN graph.}

\item{\code{clusters}}{A vector holding the cluster labels for each geneset.}

\item{\code{coregenes}}{A list of containing vectors of core genes.}

\item{\code{scores}}{A dataframe or list containg the geneset testing results.}
}}

