% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{BOWER-class}
\alias{BOWER-class}
\alias{BOWER}
\title{The BOWER container class}
\description{
The BOWER container class
}
\section{Slots}{

\describe{
\item{\code{genesets}}{A list of containing vectors of genes.}

\item{\code{graph}}{An igraph object that represents the kNN graph.}
}}

