% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_source_dir}
\alias{data_source_dir}
\title{Return the local directory of each data source in a configuration}
\usage{
data_source_dir(config)
}
\arguments{
\item{config}{data.frame: configuration as returned by \code{\link{bb_config}}}
}
\value{
character vector of directories
}
\description{
Files from each data source are stored locally in the associated directory.
}
\examples{
\dontrun{
  cf <- bb_config("/my/file/root") \%>\%
    add(bb_sources(data_group="Sea surface temperature"))
  data_source_dir(cf)
}

}
