% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthdata_handler.R
\name{earthdata_get}
\alias{earthdata_get}
\title{Handler for files files served up through Earthdata providers}
\usage{
earthdata_get(data_source)
}
\arguments{
\item{data_source}{tibble: single-row tibble defining a data source, e.g. as returned by \code{bb_source}}
}
\value{
TRUE on success
}
\description{
Handler for files files served up through Earthdata providers
}
\references{
https://wiki.earthdata.nasa.gov/display/EL/How+To+Register+With+Earthdata+Login
}
