% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_core.R, R/generics.R
\name{extract_core}
\alias{extract_core}
\alias{extract_core.BOWER}
\title{extract_core}
\usage{
\method{extract_core}{BOWER}(bower, inplace = TRUE)

extract_core(x, ...)
}
\arguments{
\item{bower}{processed bower object.}

\item{inplace}{whether or not to return an updated BOWER class or return the output as a list.}

\item{x}{a BOWER class.}

\item{...}{Arguments passed to other methods.}
}
\value{
Returns a list of genes that interect geneset clusters in coregenes slot or as a list.

Return leiden clustering assignment for each geneset.
}
\description{
extract_core
}
