% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_core.R
\name{extract_core}
\alias{extract_core}
\alias{extract_core.BOWER}
\title{extract_core}
\usage{
\method{extract_core}{BOWER}(bower, inplace = TRUE)
}
\arguments{
\item{bower}{processed bower object.}

\item{inplace}{whether or not to return an updated BOWER class or return the output as a list.}
}
\value{
Returns a list of genes that outersect geneset clusters in coregenes slot or as a list.
}
\description{
Summarize the outersection genes for the. summary terms.
}
\examples{
gmt_file <- system.file("extdata", "h.all.v7.4.symbols.gmt", package = "bowerbird")
bwr <- bower(gmt_file)
bwr <- snn_graph(bwr)
bwr <- find_clusters(bwr)
bwr <- summarize_clusters(bwr)
extract_core(bwr)
}
