% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{FAME-class}
\alias{FAME-class}
\alias{FAME}
\title{The FAME container class}
\description{
The FAME container class
}
\section{Slots}{

\describe{
\item{\code{genesets}}{A list of containing vectors of genes.}

\item{\code{graph}}{An igraph object that represents the kNN graph.}
}}

