% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R, R/generics.R
\name{find_clusters}
\alias{find_clusters}
\alias{find_clusters.BOWER}
\alias{find_clusters.igraph}
\title{clusters}
\usage{
\method{find_clusters}{BOWER}(bower, resolution = 3, ...)

\method{find_clusters}{igraph}(gr, resolution = 3, ...)

find_clusters(x, ...)
}
\arguments{
\item{bower}{BOWER class..}

\item{resolution}{value for leiden clustering.}

\item{...}{Arguments passed to other methods.}

\item{gr}{igraph object.}

\item{x}{either a BOWER class or an igraph object.}
}
\value{
Cluster assignment for each geneset.

Return leiden clustering assignment for each geneset.
}
\description{
Performs clustering of SNN graph of genesets.
}
\examples{
\donttest{
result <- find_clusters(object)
}
}
