% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghrsst.R
\name{ghrsst_get}
\alias{ghrsst_get}
\title{Handler for GHRSST data sources}
\usage{
ghrsst_get(data_source, local_dir_only = TRUE)
}
\arguments{
\item{data_source}{tibble: single-row tibble defining a data source, e.g. as returned by \code{bb_source}}

\item{local_dir_only}{logical: if TRUE, just return the local directory into which files from this data source would be saved}
}
\value{
TRUE on success
}
\description{
Handler for GHRSST data sources
}
\references{
https://podaac.jpl.nasa.gov/Multi-scale_Ultra-high_Resolution_MUR-SST
}
