% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wget.R
\name{install_wget}
\alias{install_wget}
\title{Helper function to install wget on Windows}
\usage{
install_wget()
}
\value{
TRUE (invisibly) on success
}
\description{
The wget.exe executable will be downloaded from https://eternallybored.org/misc/wget/current/wget.exe and installed into your appdata directory (typically something like C:/Users/username/AppData/Roaming/)
}
\examples{
\dontrun{
  install_wget()
}

}
\references{
https://eternallybored.org/misc/wget/current/wget.exe
}
