% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceandata.R
\name{oceandata_get}
\alias{oceandata_get}
\title{Handler for oceandata data sources}
\usage{
oceandata_get(data_source, local_dir_only = FALSE)
}
\arguments{
\item{data_source}{tibble: single-row tibble defining a data source, e.g. as returned by \code{bb_source}}

\item{local_dir_only}{logical: if TRUE, just return the local directory into which files from this data source would be saved}
}
\value{
the directory if local_dir_only is TRUE, otherwise TRUE on success
}
\description{
Handler for oceandata data sources
}
\references{
https://oceandata.sci.gsfc.nasa.gov/
}
