% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceandata.R
\name{oceandata_get}
\alias{oceandata_get}
\title{Handler for oceandata data sources}
\usage{
oceandata_get(config, verbose = FALSE, local_dir_only = FALSE)
}
\arguments{
\item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source}

\item{verbose}{logical: if TRUE, provide additional progress output}

\item{local_dir_only}{logical: if TRUE, just return the local directory into which files from this data source would be saved}
}
\value{
the directory if local_dir_only is TRUE, otherwise TRUE on success
}
\description{
Handler for oceandata data sources
}
\references{
https://oceandata.sci.gsfc.nasa.gov/
}
