% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{pp_cleanup}
\alias{pp_cleanup}
\title{Postprocessing: remove unwanted files}
\usage{
pp_cleanup(config, pattern, recursive = FALSE, ignore_case = FALSE, ...)
}
\arguments{
\item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source}

\item{pattern}{string: regular expression, passed to \code{file.info}}

\item{recursive}{logical: should the cleanup recurse into subdirectories?}

\item{ignore_case}{logical: should pattern matching be case-insensitive?}

\item{...}{: additional arguments (currently ignored)}
}
\value{
TRUE on success
}
\description{
This function is not intended to be called directly, but instead can be specified as a postprocessing step to apply to a data source.
}
\examples{
\dontrun{
  ## remove .asc files after synchronisation
  my_source <- bb_source(...,postprocess=quote(pp_cleanup,pattern="\\\\.asc$"))
}

}
\seealso{
\code{\link{bb_source}} \code{\link{bb_config}} \code{\link{pp_decompress}}
}
