% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{pp_decompress}
\alias{pp_decompress}
\alias{pp_unzip}
\alias{pp_gunzip}
\alias{pp_bunzip2}
\alias{pp_uncompress}
\title{Postprocessing: decompress zip, gz, bz2, Z files and optionally delete the compressed copy
\code{pp_unzip}, \code{pp_gunzip}, \code{pp_bunzip2}, and \code{pp_uncompress} are convenience wrappers around \code{pp_decompress} that specify the method.
The dots argument indicates additional arguments that are passed to \code{pp_decompress}. Some may be passed by \code{bb_sync}. These include parameters named \code{file_list_before} and \code{file_list_after}, which are data.frames as returned by \code{file.info}, listing the files present in the target directory before and after synchronising. These are used if delete=TRUE.}
\usage{
pp_decompress(data_source, delete = FALSE, method, ...)

pp_unzip(...)

pp_gunzip(...)

pp_bunzip2(...)

pp_uncompress(...)
}
\arguments{
\item{data_source}{tibble: single-row tibble defining a data source, e.g. as returned by \code{bb_source}}

\item{delete}{logical: delete the zip files after extracting their contents?}

\item{method}{string: one of "unzip","gunzip","bunzip2","decompress"}

\item{...}{: additional arguments passed to \code{pp_decompress}}
}
\value{
TRUE on success
}
\description{
Postprocessing: decompress zip, gz, bz2, Z files and optionally delete the compressed copy
\code{pp_unzip}, \code{pp_gunzip}, \code{pp_bunzip2}, and \code{pp_uncompress} are convenience wrappers around \code{pp_decompress} that specify the method.
The dots argument indicates additional arguments that are passed to \code{pp_decompress}. Some may be passed by \code{bb_sync}. These include parameters named \code{file_list_before} and \code{file_list_after}, which are data.frames as returned by \code{file.info}, listing the files present in the target directory before and after synchronising. These are used if delete=TRUE.
}
\examples{
\dontrun{
  ## decompress .zip files after synchronisation but keep zip files intact
  my_source <- bb_source(...,postprocess=quote(pp_unzip))

  ## decompress .zip files after synchronisation and delete zip files
  my_source <- bb_source(...,postprocess=quote(pp_unzip(delete=TRUE)))
}

}
\seealso{
\code{\link{bb_source}} \code{\link{bb_config}} \code{\link{pp_cleanup}}
}
