% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R, R/utilities.R
\docType{class}
\name{scFA-class}
\alias{scFA-class}
\alias{scFA}
\alias{init_scFA}
\title{The scFAN container class}
\usage{
init_scFA()
}
\description{
The scFAN container class
}
\section{Slots}{

\describe{
\item{\code{genesets}}{A list of containing vectors of genes.}

\item{\code{graph}}{An igraph object that represents the kNN graph.}
}}

