% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R, R/generics.R
\name{set_clusters}
\alias{set_clusters}
\alias{set_clusters.BOWER}
\title{Set cluster to BOWER class}
\usage{
\method{set_clusters}{BOWER}(bower, clusters)

set_clusters(x, ...)
}
\arguments{
\item{bower}{bower class}

\item{clusters}{Cluster assignment for each geneset.}

\item{x}{BOWER class.}

\item{...}{Arguments passed to other methods.}
}
\value{
bower class

Set cluster labels for each gene set
}
\description{
Set cluster to BOWER class
}
\details{
Manually sets externally/alternatively determined cluster assignment for each geneset.
}
\examples{
\donttest{
bower <- set_clusters(bower, c(1,1,1,2,2,2,3,3,3))
}
}
