% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{set_clusters}
\alias{set_clusters}
\alias{set_clusters.BOWER}
\title{Set cluster to BOWER class}
\usage{
\method{set_clusters}{BOWER}(bower, clusters)
}
\arguments{
\item{bower}{bower class}

\item{clusters}{Cluster assignment for each geneset.}
}
\value{
bower class
}
\description{
Manually sets externally/alternatively determined cluster assignment for each geneset.
}
\examples{
gmt_file <- system.file("extdata", "h.all.v7.4.symbols.gmt", package = "bowerbird")
bwr <- bower(gmt_file)
bwr <- set_clusters(bwr, rep(1:5, each = 10)) # e.g. randomly make 5 blocks of 10 clusters
}
