% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wget.R
\name{wget}
\alias{wget}
\title{Use wget to mirror an external data repository}
\usage{
wget(url, flags)
}
\arguments{
\item{url}{string: the URL to retrieve}

\item{flags}{string: command-line flags to pass to wget}
}
\value{
TRUE on success
}
\description{
Use wget to mirror an external data repository
}
\seealso{
\code{\link{install_wget}}
}
