% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wget.R
\name{wget}
\alias{wget}
\title{Make a wget call}
\usage{
wget(url, flags, verbose = TRUE)
}
\arguments{
\item{url}{string: the URL to retrieve}

\item{flags}{string: command-line flags to pass to wget}

\item{verbose}{logical: print trace output?}
}
\value{
TRUE on success
}
\description{
Make a wget call
}
\seealso{
\code{\link{install_wget}}
}
