% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wget.R
\name{wget}
\alias{wget}
\title{Make a wget call}
\usage{
wget(url, flags, verbose = FALSE, ...)
}
\arguments{
\item{url}{string: the URL to retrieve}

\item{flags}{string: command-line flags to pass to wget}

\item{verbose}{logical: print trace output?}

\item{...}{: additional paramaters passed to \code{system2}}
}
\value{
the result of the system2 call
}
\description{
Make a wget call
}
\examples{
\dontrun{
## get help about wget command line parameters
wget("--help")
}

}
\seealso{
\code{\link{install_wget}} \code{\link{system2}}
}
